moved the BaseTerminal class into the platform header
