+++ cpp-terminal/base.hpp
@@ -2,6 +2,8 @@
 
 #include <string>
 #include <cpp-terminal/terminal_base.h>
+// TODO: remove include
+#include "private/platform.hpp"
 
 namespace Term {
 
@@ -109,7 +111,7 @@ void get_term_size_slow(int&, int&);
 
 
 // initializes the terminal
-class Terminal : public BaseTerminal {
+class Terminal : public Private::BaseTerminal {
    public:
     Terminal(bool enable_keyboard, bool disable_ctrl_c);
     // providing no parameters will disable the keyboard and ctrl+c
